backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
email-validator
reportlab
passlib[bcrypt]
